+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Patch 11.0.0, updating dependencies
 -   Taint Fix: you are not suppose to move stuff while in combat.
+-   Fix minimap tracking frame.
 
 ## [1.0.0] - 2024-07-27